requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=1.4
asyncpg
alembic
python-dotenv
passlib[bcrypt]
python-jose 